backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# Task Queue
celery==5.4.0
redis==5.2.1

# Meta Ads
facebook-business==21.0.2

# HTTP Client
httpx==0.28.1

# Security
cryptography==44.0.0
pyjwt==2.9.0
bcrypt==4.2.0

# Form parsing
python-multipart==0.0.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Logging
python-dotenv==1.0.1
